Basic UI complete
Note:
The design does not look good at all, but focus on the code for now and work on the design while coding if any idea pops up, otherwise keep it like this until everything is done.

Next:
  i. Write the basic class for the view controllers.
 ii. Handle the key (DM5 encryption).
iii. Fetch and parse the data of the characters.
iv. Fetch and parse the data of the comics.

A plus: Clean the code and look for performance improvements.